Create proverbs
The program helps the user memorise proverbial phrases. It randomly shows an entry of a dictionary of incomplete phrases for the user to complete. When the user enters a words, the program checks to see if correct. If so, the program congratulates the user. Otherwise, the program tells the user they are incorrect and displays the phrase with the missing word filled in.